fixed schema import + refactored update state to be in update workflow
